fix overflow when reading zarr at large indices

+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/precomputed/PrecomputedHeader.scala
@@ -74,18 +74,21 @@ case class PrecomputedScaleHeader(precomputedScale: PrecomputedScale, precompute
 
   lazy val compressorImpl: Compressor = PrecomputedCompressorFactory.create(this)
 
-  override def chunkShapeAtIndex(chunkIndex: Array[Int]): Array[Int] =
+  override def chunkShapeAtIndex(chunkIndex: Array[Long]): Array[Int] =
     chunkIndexToNDimensionalBoundingBox(chunkIndex).map(dim => dim._2 - dim._1)
 
   override def voxelOffset: Array[Int] = precomputedScale.voxel_offset.getOrElse(Array(0, 0, 0))
 
-  def chunkIndexToNDimensionalBoundingBox(chunkIndex: Array[Int]): Array[(Int, Int)] =
+  // Neuroglancer precomputed chunk grids are bounded spatial voxel coordinates (unlike a segment-id-keyed
+  // mapping array), so narrowing back to Int here is safe -- chunkIndex is Long only because it comes from
+  // the shared DatasetArray/ChunkUtils chunk-addressing interface.
+  def chunkIndexToNDimensionalBoundingBox(chunkIndex: Array[Long]): Array[(Int, Int)] =
     chunkIndex.zipWithIndex.map { chunkIndexWithDim =>
       val (chunkIndexAtDim, dim) = chunkIndexWithDim
       val beginOffset = voxelOffset(dim) + chunkIndexAtDim * precomputedScale.primaryChunkShape(dim)
       val endOffset = voxelOffset(dim) + ((chunkIndexAtDim + 1) * precomputedScale.primaryChunkShape(dim))
-        .min(precomputedScale.size(dim).toInt)
-      (beginOffset, endOffset)
+        .min(precomputedScale.size(dim))
+      (beginOffset.toInt, endOffset.toInt)
     }
 
   def gridSize: Array[Int] = chunkShape.zip(precomputedScale.size).map { case (c, s) => (s.toDouble / c).ceil.toInt }
